fixed issue: multiple I commands
Fixed issue where multiple I commands at the same position (in different
sequences then), would not be printed properly.
Fixed by making the portion of code that deals with I commands to
behave the same as other commands: dealing with one character at a time.
This is the current method implemented in the naive_parser() function.